+++ packages/sandbox-local/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @effect-agent/sandbox-local
 
+## 0.1.0-beta.72
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/sandbox@0.1.0-beta.72
+
+## 0.1.0-beta.71
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/sandbox@0.1.0-beta.71
+
 ## 0.1.0-beta.70
 
 ### Patch Changes