+++ packages/sandbox/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @effect-agent/sandbox
 
+## 0.1.0-beta.72
+
+## 0.1.0-beta.71
+
 ## 0.1.0-beta.70
 
 ## 0.1.0-beta.69